requirements-gpu.txt
# Python 3.10
# conda create -n object-recognition-train3.10 python=3.10
# conda activate object-recognition-train3.10
# pip install -r requirements-gpu.txt -i https://pypi.tuna.tsinghua.edu.cn/simple --verbose

numpy==1.23.5
scikit-learn==1.7.0
matplotlib==3.10.3
pillow==11.3.0
tqdm==4.67.1
keras==2.12.0
tensorflow_gpu==2.12.0